move student related profile components into subfolder
