+++ src/components/profile/student/CourseOfStudySection.vue
